Reorganizes packets
