+++ Projects/UOContent/Items/Games/Mahjong/MahjongPacketHandlers.cs
@@ -25,7 +25,7 @@ namespace Server.Engines.Mahjong
 
         public static void Initialize()
         {
-            Packets.Register(0xDA, 0, true, OnPacket);
+            IncomingPackets.Register(0xDA, 0, true, OnPacket);
 
             RegisterSubCommand(0x6, ExitGame);
             RegisterSubCommand(0xA, GivePoints);
