+++ Projects/UOContent/Items/Misc/BulletinBoards.cs
@@ -229,7 +229,7 @@ namespace Server.Items
 
         public static void Initialize()
         {
-            Packets.Register(0x71, 0, true, BBClientRequest);
+            IncomingPackets.Register(0x71, 0, true, BBClientRequest);
         }
 
         public static void BBClientRequest(NetState state, CircularBufferReader reader)
